update migration file

+++ database/migrations/2022_09_24_200437_save_judgement.php
@@ -15,13 +15,16 @@ class SaveJudgement extends Migration
      */
     public function up()
     {
-        Schema::table('submissions', function (Blueprint $table) {
-            //
-            $table->json('judgement');
-        });
+        if (!Schema::hasColumn('submissions')){
+
+            Schema::table('submissions', function (Blueprint $table) {
+                //
+                $table->json('judgement');
+            });
+        }
 
         $progress = 0;
-        Submission::chunk(400, function($subs) use ($progress){
+        Submission::chunk(400, function($subs) use (&$progress){
             foreach ($subs as  $sub){
                 $result = file_get_contents( $sub->directory() . "/result-" . $sub->id . ".html");
                 $results = explode("</span>\n", $result);
@@ -61,9 +64,12 @@ class SaveJudgement extends Migration
      */
     public function down()
     {
-        Schema::table('submissions', function (Blueprint $table) {
-            //
-            $table->dropColumn('judgement');
-        });
+        if (Schema::hasColumn('submissions')){
+
+            Schema::table('submissions', function (Blueprint $table) {
+                //
+                $table->dropColumn('judgement');
+            });
+        }
     }
 }
